fix(remoter): 重构 AgentModelProvider 类
